+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add `segment_text` function for breaking text into sentences with character-level indices
 - Add `corpus` submodule with input classes for text and TEI/XML and factory
   method to initialize appropriate input class based on file type
-- Processes TEI/XML documents to yield separate chunks for body text and footnotes, with each chunk labeled by section type
+- Processes TEI/XML documents to yield separate chunks for body text and footnotes, with each footnote yielded individually as a separate element
 
 ### Application